backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# OCR and PDF processing
pytesseract==0.3.10
pdf2image==1.17.0
PyMuPDF==1.23.8
Pillow==10.2.0
Poppler==24.08.0
# Google Gemini AI
google-generativeai==0.3.2

# Google Sheets integration (for Phase 3)
gspread==5.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
